64bit: bot_weaponstate_t side-band + LibVarValue float ABI + aasworld.entities typed + dword_100643A8 char* + chatdumps side-band
- bot_weaponstate_t side-band botweaponstates[MAXCLIENTS]:
  retypes bs->weaponweights pointer slots without breaking the
  embedded int[7] layout used by 32-bit Windows DLL. NULL-safe
  sub_10035640 so BotResetState can run before BotSetupClient.

- botlib_exports.c LibVarValue extern prototype fixed:
  was declared double LibVarValue(char*, int) — real signature is
  float LibVarValue(char*, char*). aarch64 returns float in s0 and
  double in d0 so the wrong proto silently returned junk; this is

- aas_world_t::entities retyped from int to aas_entity_t*; all
  ~10 byte-arithmetic sites wrapped with (char *) casts.

- dword_100643A8 (per-client settings buffer) retyped from int to
  char *; all 5 byte-arithmetic call sites cleaned up.

  truncating int pointer.

- chatstate +184 pointer slot routed through new side-band
  botchatdumps[MAXCLIENTS] via BotChatDumpSlot(cs) macro.

- BotDumpInitialChat v13 retyped to source_t * (was int truncating
  LoadSourceFile return).

Still broken: chat-dump heap structure (BotDumpInitialChat output)
pointers.

Source-commit: 2b8233fc61e0a460e82a4c76155baaf8f1f83f83

+++ botlib/botlib_exports.c
@@ -61,7 +61,7 @@ extern int  maxentities;     /* @ 0x10064024 */
 /* Slot 1 helpers */
 extern int    Swap_Init(void);                            /* 0x100439F0 */
 extern int    BotSetupLibrary(void);                      /* 0x10029C90 inner */
-extern double LibVarValue(char *name, char *default_str); /* 0x10038A90 */
+extern float  LibVarValue(char *name, char *default_str); /* 0x10038A90 */
 extern int    BotSetupMoveAI(void);                       /* 0x10037A00 */
 extern int    sub_1000EDC0(int maxentities, int maxclients);
 extern int    EA_Setup(void);                             /* 0x10037660 */
